Add initial build script
And tweak versions of libraries to avoid problem building gcc.

No idea why this is required.

It looks like this problem from 2014:

https://stackoverflow.com/questions/23128949/compiling-gcc-cannot-find-output-from-flex-giving-up

and one of the comments suggested that later versions of the libraries
could be used, so that's what was tried and seemed to work OK to build
Rockbox for the Sansa e200.